+++ src/main/kotlin/xyz/cssxsh/bilibili/data/user/UserSimple.kt
@@ -2,14 +2,13 @@ package xyz.cssxsh.bilibili.data.user
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.JsonObject
 
 @Serializable
 data class UserSimple(
     @SerialName("face")
     val face: String? = null,
     @SerialName("uid")
-    val uid: Int,
+    val uid: Long,
     @SerialName("uname")
     val uname: String? = null
 )
